fix(complete)!: Move Completer into shells
It only exists for shell-adaptation, so it feels like it better belongs
there.
