--- clap_complete/src/dynamic/completer.rs
@@ -1,23 +1,0 @@
-use std::ffi::OsString;
-
-/// Shell-specific completions
-pub trait Completer {
-    /// The recommended file name for the registration code
-    fn file_name(&self, name: &str) -> String;
-    /// Register for completions
-    fn write_registration(
-        &self,
-        name: &str,
-        bin: &str,
-        completer: &str,
-        buf: &mut dyn std::io::Write,
-    ) -> Result<(), std::io::Error>;
-    /// Complete the given command
-    fn write_complete(
-        &self,
-        cmd: &mut clap::Command,
-        args: Vec<OsString>,
-        current_dir: Option<&std::path::Path>,
-        buf: &mut dyn std::io::Write,
-    ) -> Result<(), std::io::Error>;
-}
