+++ clap_complete/src/dynamic/mod.rs
@@ -4,13 +4,11 @@
 
 mod candidate;
 mod complete;
-mod completer;
 mod custom;
 
 pub mod shells;
 
 pub use candidate::CompletionCandidate;
 pub use complete::complete;
-pub use completer::Completer;
 pub use custom::ArgValueCompleter;
 pub use custom::CustomCompleter;
