+++ clap_complete/src/dynamic/shells/zsh.rs
@@ -2,7 +2,7 @@
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
 pub struct Zsh;
 
-impl crate::dynamic::Completer for Zsh {
+impl crate::dynamic::shells::Completer for Zsh {
     fn file_name(&self, name: &str) -> String {
         format!("{name}.zsh")
     }
